requirements-streamlit.txt
# ─── Streamlit Requirements for Streamlit Cloud (1GB RAM free tier) ───────────
# This is for the Streamlit version of the app

streamlit>=1.28.0

# ─── LLM Integration ─────────────────────────────────────────────────────────
groq>=0.9.0               # Groq SDK — FREE API (llama-3, mixtral, gemma)
openai>=1.35.3            # OpenAI-compatible SDK — for xAI or OpenAI GPT

# ─── Web Search & HTTP ───────────────────────────────────────────────────────
duckduckgo-search>=6.2.0  # Zero-key web search for real-time results
httpx>=0.27.0             # HTTP client for URL summarization

# ─── Utilities ───────────────────────────────────────────────────────────────
python-dotenv>=1.0.1
